user-python/requirements.txt
anyio==3.5.0
asgiref==3.5.1
asyncpg==0.25.0
atlastk==0.13
click==8.1.3
Deprecated==1.2.13
distconfig3==1.0.1
fastapi==0.75.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
h11==0.13.0
idna==3.3
iso8601==1.0.2
opentelemetry-api==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-exporter-jaeger-proto-grpc==1.11.1
opentelemetry-exporter-jaeger-thrift==1.11.1
opentelemetry-instrumentation==0.30b1
opentelemetry-instrumentation-asgi==0.30b1
opentelemetry-instrumentation-asyncpg==0.30b1
opentelemetry-instrumentation-fastapi==0.30b1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
opentelemetry-util-http==0.30b1
protobuf==3.20.1
pydantic==1.9.0
PyPika==0.48.9
pytz==2022.1
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
structlog==21.5.0
thrift==0.16.0
toml==0.10.2
typing_extensions==4.2.0
uvicorn==0.17.6
vyper-config==1.1.0
watchdog==2.1.7
wrapt==1.14.0
